PHP folder, Contact Form to Database
Added a folder for Contact with a controller page (index.php,
contact.html.php, error and success pages. The controller page
successfully writes to the DB using Xampp. Deleted unneeded fields in
the forms for register and contact. Fixed the gap in the content for
the home page paragraphs by adding them to one section. Minor fixes in
the nav to accommodate the new contact folder.