Weather API, Added Anton Google Font, Forms path
Added weather API to the Home, Register, and Contact Page. Added the ID
associated with the weather API to header includes. Added a font for
the logo and the weather api called Anton from Google fonts. Fixed the
path for the forms on the home page as I didn't see them in mobile.
Works now.